requirements.txt
fastapi
uvicorn
replicate
python-dotenv
pydantic
web3